BgCombine/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.7.0
aiofiles=24.1.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.13.2=pypi_0
aiosignal=1.4.0=pypi_0
annotated-doc=0.0.4=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.12.0=pypi_0
astor=0.8.1=pypi_0
asttokens=3.0.1=pypi_0
attrs=25.4.0=pypi_0
blinker=1.9.0=pypi_0
brotli=1.2.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.12.2=haa95532_0
ccimport=0.4.4=pypi_0
certifi=2025.11.12=pypi_0
charset-normalizer=3.4.4=pypi_0
click=8.3.1=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.3=pypi_0
configargparse=1.7.1=pypi_0
contourpy=1.3.3=pypi_0
cumm-cu121=0.7.11=pypi_0
cycler=0.12.1=pypi_0
cython=3.2.3=pypi_0
dash=3.3.0=pypi_0
dataclasses-json=0.6.7=pypi_0
debugpy=1.8.19=pypi_0
decorator=5.2.1=pypi_0
deprecated=1.3.1=pypi_0
easydict=1.13=pypi_0
einops=0.8.1=pypi_0
entrypoints=0.4=pypi_0
executing=2.2.1=pypi_0
expat=2.7.3=h885b0b7_4
fastapi=0.124.4=pypi_0
fastcore=1.9.2=pypi_0
fastjsonschema=2.21.2=pypi_0
ffmpy=1.0.0=pypi_0
filelock=3.20.0=pypi_0
findpydeps=0.2.8=pypi_0
fire=0.7.1=pypi_0
flash-attn=2.8.3+pt2.5.0cu121=pypi_0
flask=3.1.2=pypi_0
fonttools=4.61.1=pypi_0
freetype-py=2.5.1=pypi_0
frozenlist=1.8.0=pypi_0
fsspec=2025.12.0=pypi_0
glcontext=3.0.0=pypi_0
gradio=6.1.0=pypi_0
gradio-client=2.0.1=pypi_0
greenlet=3.3.0=pypi_0
groovy=0.1.2=pypi_0
gsplat=1.5.3=pypi_0
h11=0.16.0=pypi_0
hf-xet=1.2.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=1.2.3=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.11=pypi_0
igraph=1.0.0=pypi_0
imageio=2.33.0=pypi_0
importlib-metadata=8.7.0=pypi_0
iniconfig=2.3.0=pypi_0
iopath=0.1.10=pypi_0
ipycanvas=0.14.3=pypi_0
ipyevents=2.0.4=pypi_0
ipykernel=6.29.5=pypi_0
ipython=9.8.0=pypi_0
ipython-pygments-lexers=1.1.1=pypi_0
ipywidgets=8.1.8=pypi_0
itsdangerous=2.2.0=pypi_0
jaxtyping=0.3.4=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
jsonschema=4.25.1=pypi_0
jsonschema-specifications=2025.9.1=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-core=5.9.1=pypi_0
jupyterlab-widgets=3.0.16=pypi_0
kaolin=0.18.0=pypi_0
kiwisolver=1.4.9=pypi_0
lark=1.3.1=pypi_0
libexpat=2.7.3=h885b0b7_4
libffi=3.4.4=hd77b12b_1
libzlib=1.3.1=h02ab6af_0
lightning=2.6.0=pypi_0
lightning-utilities=0.15.2=pypi_0
loguru=0.7.3=pypi_0
lovely-numpy=0.2.18=pypi_0
lovely-tensors=0.1.20=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.26.1=pypi_0
matplotlib=3.10.8=pypi_0
matplotlib-inline=0.2.1=pypi_0
mdurl=0.1.2=pypi_0
moderngl=5.12.0=pypi_0
moge=1.0.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.7.0=pypi_0
mypy-extensions=1.1.0=pypi_0
narwhals=2.14.0=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.6.1=pypi_0
ninja=1.13.0=pypi_0
numpy=2.2.6=pypi_0
omegaconf=2.3.0=pypi_0
open3d=0.19.0=pypi_0
opencv-python=4.12.0.88=pypi_0
openssl=3.0.18=h543e019_0
optree=0.18.0=pypi_0
orjson=3.11.5=pypi_0
packaging=25.0=pypi_0
pandas=2.3.3=pypi_0
parso=0.8.5=pypi_0
pccm=0.4.16=pypi_0
pillow=12.0.0=pypi_0
pip=25.3=pyhc872135_0
pipdeptree=2.30.0=pypi_0
platformdirs=4.5.1=pypi_0
playwright=1.57.0=pypi_0
plotly=6.5.0=pypi_0
pluggy=1.6.0=pypi_0
plyfile=1.1.3=pypi_0
pooch=1.8.2=pypi_0
portalocker=3.2.0=pypi_0
prompt-toolkit=3.0.52=pypi_0
propcache=0.4.1=pypi_0
psutil=7.1.3=pypi_0
pure-eval=0.2.3=pypi_0
pybind11=3.0.1=pypi_0
pydantic=2.12.4=pypi_0
pydantic-core=2.41.5=pypi_0
pydub=0.25.1=pypi_0
pyee=13.0.0=pypi_0
pyglet=2.1.11=pypi_0
pygltflib=1.16.5=pypi_0
pygments=2.19.2=pypi_0
pymeshfix=0.17.1=pypi_0
pyopengl=3.1.0=pypi_0
pyopengl-accelerate=3.1.10=pypi_0
pyparsing=3.2.5=pypi_0
pyqt6=6.10.1=pypi_0
pyqt6-qt6=6.10.1=pypi_0
pyqt6-sip=13.10.3=pypi_0
pyqtgraph=0.14.0=pypi_0
pyrender=0.1.45=pypi_0
pytest=9.0.2=pypi_0
python=3.11.14=h981015d_0
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.20=pypi_0
pytorch-lightning=2.6.0=pypi_0
pytorch3d=0.7.8+5043d15pt2.5.1cu121=pypi_0
pytz=2025.2=pypi_0
pyvista=0.46.4=pypi_0
pywin32=311=pypi_0
pyyaml=6.0.3=pypi_0
pyzmq=27.1.0=pypi_0
referencing=0.37.0=pypi_0
requests=2.32.5=pypi_0
retrying=1.4.2=pypi_0
rich=14.2.0=pypi_0
rpds-py=0.30.0=pypi_0
safehttpx=0.1.7=pypi_0
safetensors=0.7.0=pypi_0
sam3d-objects=0.0.1=pypi_0
scipy=1.16.3=pypi_0
scooby=0.11.0=pypi_0
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=80.9.0=py311haa95532_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
spconv-cu121=2.3.8=pypi_0
sqlite=3.51.0=hda9a48d_0
stack-data=0.6.3=pypi_0
starlette=0.50.0=pypi_0
sympy=1.13.1=pypi_0
termcolor=3.2.0=pypi_0
texttable=1.7.0=pypi_0
timm=1.0.22=pypi_0
tk=8.6.15=hf199647_0
tomlkit=0.13.3=pypi_0
torch=2.5.1+cu121=pypi_0
torchaudio=2.5.1+cu121=pypi_0
torchmetrics=1.8.2=pypi_0
torchvision=0.20.1+cu121=pypi_0
tornado=6.5.4=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
trimesh=4.10.0=pypi_0
triton-windows=3.1.0.post17=pypi_0
typer=0.20.0=pypi_0
typer-slim=0.20.0=pypi_0
typing-extensions=4.15.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing-inspection=0.4.2=pypi_0
tzdata=2025.3=pypi_0
ucrt=10.0.22621.0=haa95532_0
urllib3=2.6.2=pypi_0
usd-core=25.11=pypi_0
utils3d=0.0.2=pypi_0
uvicorn=0.38.0=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
vtk=9.5.2=pypi_0
wadler-lindig=0.1.7=pypi_0
warp-lang=1.10.1=pypi_0
wcwidth=0.2.14=pypi_0
werkzeug=3.1.4=pypi_0
wheel=0.45.1=py311haa95532_0
widgetsnbextension=4.0.15=pypi_0
win32-setctime=1.2.0=pypi_0
wrapt=2.0.1=pypi_0
xatlas=0.0.11=pypi_0
xformers=0.0.28.post3=pypi_0
xz=5.6.4=h4754444_1
yarl=1.22.0=pypi_0
zipp=3.23.0=pypi_0
zlib=1.3.1=h02ab6af_0
zstandard=0.25.0=pypi_0
